--- public/images/placeholder-avatar-1.jpg
@@ -1,1 +1,0 @@
-data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAiIGhlaWdodD0iNDAiIHZpZXdCb3g9IjAgMCA0MCA0MCIgZmlsbD0ibm9uZSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj4KPHJlY3Qgd2lkdGg9IjQwIiBoZWlnaHQ9IjQwIiByeD0iMjAiIGZpbGw9IiNFNUU3RUIiLz4KPHBhdGggZD0iTTIwIDE1QzIyLjA3NjEgMTUgMjMuNzUgMTYuNjczOSAyMy43NSAxOC43NUMyMy43NSAyMC44MjYxIDIyLjA3NjEgMjIuNSAyMCAyMi41QzE3LjkyMzkgMjIuNSAxNi4yNSAyMC44MjYxIDE2LjI1IDE4Ljc1QzE2LjI1IDE2LjY3MzkgMTcuOTIzOSAxNSAyMCAxNVoiIGZpbGw9IiM5Q0EzQUYiLz4KPHBhdGggZD0iTTEyLjUgMjcuNUMxMi41IDI0LjUgMTUuNSAyMi41IDIwIDIyLjVDMjQuNSAyMi41IDI3LjUgMjQuNSAyNy41IDI3LjVIMTguNzVDMTguNzUgMjcuNSAxNy44NzUgMjcuNSAxNi41MjUgMjcuNUMxNS4xNzUgMjcuNSAxMy41IDI3LjUgMTIuNSAyNy41WiIgZmlsbD0iIzlDQTNBRiIvPgo8L3N2Zz4K
