--- public/images/placeholder-avatar-4.jpg
@@ -1,1 +1,0 @@
-data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAiIGhlaWdodD0iNDAiIHZpZXdCb3g9IjAgMCA0MCA0MCIgZmlsbD0ibm9uZSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj4KPHJlY3Qgd2lkdGg9IjQwIiBoZWlnaHQ9IjQwIiByeD0iMjAiIGZpbGw9IiNFMEY3RkUiLz4KPHBhdGggZD0iTTIwIDE1QzIyLjA3NjEgMTUgMjMuNzUgMTYuNjczOSAyMy43NSAxOC43NUMyMy43NSAyMC44MjYxIDIyLjA3NjEgMjIuNSAyMCAyLjVDMTcuOTIzOSAyMi41IDE2LjI1IDIwLjgyNjEgMTYuMjUgMTguNzVDMTYuMjUgMTYuNjczOSAxNy45MjM5IDE1IDIwIDE1WiIgZmlsbD0iIzA2YjZkMCIvPgo8cGF0aCBkPSJNMTIuNSAyNy41QzEyLjUgMjQuNSAxNS41IDIyLjUgMjAgMjIuNUMyNC41IDIyLjUgMjcuNSAyNC41IDI3LjUgMjcuNUgxOC43NUMxOC43NSAyNy41IDE3Ljg3NSAyNy41IDE2LjUyNSAyNy41QzE1LjE3NSAyNy41IDEzLjUgMjcuNSAxMi41IDI3LjVaIiBmaWxsPSIjMDZiNmQwIi8+Cjwvc3ZnPgo=
